requirements.txt
requests
mysql-connector-python
gql
aiohttp
python-dotenv 